Added changes to UI of collection navigation.
