+++ src/components/images/SingleImage.tsx
@@ -1,22 +1,24 @@
-import { Box } from "@chakra-ui/react";
+import { Box, Tooltip } from "@chakra-ui/react";
+import { useNavigate } from "react-router-dom";
 
 type ImageCollection = {
   image: {
     id: number;
     name: string;
   };
-  childToParent: any;
 };
 
 const SingleImage = (prop: ImageCollection) => {
+  const navigation = useNavigate();
+  const showSingleImageDetails = () => {
+    navigation(`/collection/${prop.image.id}`);
+  };
   return (
-    <Box
-      cursor={"pointer"}
-      p={2}
-      onClick={() => prop.childToParent(prop.image)}
-    >
-      {prop.image.name}
-    </Box>
+    <Tooltip label="Click here to see collection!">
+      <Box cursor={"pointer"} p={2} onClick={showSingleImageDetails}>
+        {prop.image.name}
+      </Box>
+    </Tooltip>
   );
 };
