+++ src/routes.ts
@@ -1,4 +1,5 @@
 import Home from "./components/home/Home";
+import SingleImageDetails from "./components/images/SingleImageDetails";
 import Login from "./components/login/Login";
 import Registration from "./components/registration/Registration";
 
@@ -17,6 +18,10 @@ const LOGGED_IN_USER_ROUTES = [
   {
     path: "/",
     element: Home,
+  },
+  {
+    path: "/collection/:id",
+    element: SingleImageDetails,
   },
 ];
